PhaseSpaceTools.jl

PhaseSpaceTools.jl

A julia package for working with quantum phase space distributions.

Overview

This package provides sampling methods for commonly used quantum states in various quantum phase-space representations, including Glauber-P, Positive-P, HusimiQ, and Wigner distributions.

There are also convenience methods for calculating operator averages from phase-space averages, and for sampling noises for solving SDEs in DifferentialEquations.jl.

Installation

In the julia REPL, enter

julia> ]add https://github.com/AshtonSBradley/PhaseSpaceTools.jl

Pkg> test PhaseSpaceTools

Usage Examples

Coherent state

A trivial example is provided by the coherent state $|\alpha\rangle$. These are eigenstate of the Bose annihilation operator $a$, with commutation relation $[a,a^\dagger]=1$. They satisfy the eigenvalue equation

\[ a|\alpha\rangle = \alpha|\alpha\rangle\]

The Fock basis $|n\rangle$ is often convenient to work in. Fock states are eigenstates of the number operator $n=a^\dagger a$:

\[a^\dagger a|n\rangle = n|n\rangle.\]

In the Fock basis, the coherent states take the form

\[|\alpha\rangle = e^{-|\alpha|^2/2}\sum_{n=0}^\infty\frac{\alpha^n}{n!}|n\rangle.\]

Coherent states play a central role in quantum phase space methods, providing a mapping of many-body boson dynamical problems to equivalent stochastic differential equations.

Glauber-P

In the Glauber-P representation the state may be sampled as a single point on the complex plane

N = 10000
state = Coherent(12.0)
α,ᾱ = glauberP(state,N)

Positive-P

In the Positive-P representation the simples way to sample the state is again as a point on the complex plane

α,ᾱ = positiveP(state,N)

Fock state

A more difficult state to sample is the eigenstate of the number operator...

Quantum states

Quantum states available in PhaseSpaceTools.jl are

Coherent

PhaseSpaceTools.Coherent — Type.

Coherent(β)

Create a coherent state with complex amplitude β.

source

Fock

PhaseSpaceTools.Fock — Type.

Fock(n)

Create a Fock state for particle number n.

source

Crescent

PhaseSpaceTools.Crescent — Type.

Crescent(β,ϵ,q)

Create a Cresecent state with parameters β, ϵ, q.

source

Squeezed

PhaseSpaceTools.Squeezed — Type.

Squeezed(β,ϵ)

Create a Squeezed state with parameters β, ϵ.

source

Thermal

PhaseSpaceTools.Thermal — Type.

Thermal(β,n̄)

Create a Thermal state with parameters β, n̄.

source

Bogoliubov

PhaseSpaceTools.Bogoliubov — Type.

Bogoliubov(u,v,n̄)

Create a Bogoliubov state with parameters u, v, n̄.

source

Distributions

The phase-space distributions that may be sampled for these states are

Husimi-Q

PhaseSpaceTools.husimiQ — Function.

α,ᾱ = husimiQ(state <: State,N)

Generate N samples from the Husimi-Q phase-space distribution for state.

Moments of the Husimi-Q distribution generate quantum operator averages that are anti-normally ordered.

source

Glauber-P

PhaseSpaceTools.glauberP — Function.

α,ᾱ = glauberP(state <: State,N)

Generate N samples from the Glauber-Sudarshan-P phase-space distribution for state.

Moments of the Glauber-Sudarshan-P distribution generate quantum operator averages that are normally ordered.

source

Positive-P

PhaseSpaceTools.positiveP — Function.

α,ᾱ = positiveP(state <: State,N)

Generate N samples from the positive-P phase-space distribution for state.

Moments of the positive-P distribution generate quantum operator averages that are normally ordered.

source

Wigner

PhaseSpaceTools.wigner — Function.

α,ᾱ = wigner(state <: State,N)

Generate N samples from wigner phase-space distribution for state.

Moments of the Wigner distribution generate symmetrically ordered quantum operator averages.

source

Positive-W

PhaseSpaceTools.positiveW — Function.

α,ᾱ = positiveW(state <: State,N)

Generate N samples from the positive-W phase-space distribution for state.

Implemented states are

Fock(N)

The moments of the positive-W distribution generate quantum operator averages that are symmmetrically ordered.

source

Noises

randnc

PhaseSpaceTools.randnc — Function.

a = randnc(N...)

Returns an array of dimension length(N), containing samples of complex random variables with mean zero and variance one:

\[\langle |a|^2\rangle = 1; \quad\quad\langle a^2\rangle = \langle (a^*) ^2\rangle = 0,\]

Useful for creating a range of noises that show up in phase-space simulations.

Usage:

a = randnc(10) returns a 10-element Array{Complex{Float64},1}.

a = randnc(50,100) returns a 50x100-element Array{Complex{Float64},2}.

source

realnoise

PhaseSpaceTools.realnoise — Function.

Complex variable SDE's in DifferentialEquations.jl have complex noises by default dispatch on complex fields. This method defines a real noise for use with the library.

source

realbridge

PhaseSpaceTools.realbridge — Function.

Complex variable SDE's in DifferentialEquations.jl have complex noises by default dispatch on complex fields. This method defines a real brownian bridge for use with the library.

source

Solving SDEs

Complex noise example

Real noise example

Recovering normal order

A common approach to quantum phase space simulations involves performing truncated Wigner simulations, necessitating symmetric order of operator moments. We can easily recover normal order, the experimentally relevant form, using

\[\langle (a^\dagger)^p a^q\rangle = \sum_{n=0}^{\textrm{min}(p,q)}\]

Citing

Index

PhaseSpaceTools.Bogoliubov

PhaseSpaceTools.Coherent

PhaseSpaceTools.Crescent

PhaseSpaceTools.Fock

PhaseSpaceTools.Squeezed

PhaseSpaceTools.Thermal

PhaseSpaceTools.glauberP

PhaseSpaceTools.husimiQ

PhaseSpaceTools.positiveP

PhaseSpaceTools.positiveW

PhaseSpaceTools.randnc

PhaseSpaceTools.realbridge

PhaseSpaceTools.realnoise

PhaseSpaceTools.wigner