Solving SDES

Solving SDEs

Noises

randnc

PhaseSpaceTools.randnc — Function.

a = randnc(N...)

Returns an array of dimension length(N), containing samples of complex random variables with mean zero and variance one:

\[\langle |a|^2\rangle = 1; \quad\quad\langle a^2\rangle = \langle (a^*) ^2\rangle = 0,\]

Useful for creating a range of noises that show up in phase-space simulations.

Usage:

a = randnc(10) returns a 10-element Array{Complex{Float64},1}.

a = randnc(50,100) returns a 50x100-element Array{Complex{Float64},2}.

source

realnoise

PhaseSpaceTools.realnoise — Function.

Complex variable SDE's in DifferentialEquations.jl have complex noises by default dispatch on complex fields. This method defines a real noise for use with the library.

source

realbridge

PhaseSpaceTools.realbridge — Function.

Complex variable SDE's in DifferentialEquations.jl have complex noises by default dispatch on complex fields. This method defines a real brownian bridge for use with the library.

source

Complex noise example

Real noise example

Recovering normal order

A common approach to quantum phase space simulations involves performing truncated Wigner simulations, necessitating symmetric order of operator moments. We can easily recover normal order, the experimentally relevant form, using

```math \langle (a^\dagger)^p a^q\rangle = \sum_{n=0}^{\textrm{min}(p,q)}