Quantum states

Quantum states available in PhaseSpaceTools.jl are

Coherent

PhaseSpaceTools.Coherent — Type

Coherent(β)

Create a coherent state with complex amplitude β ∈ ℂ.

Fock

PhaseSpaceTools.Fock — Type

Fock(n)

Create a Fock state for particle number n ∈ ℕ₀.

Crescent

PhaseSpaceTools.Crescent — Type

Crescent(β,ϵ,q)

Create a Cresecent state with parameters β, ϵ, q.

Squeezed

PhaseSpaceTools.Squeezed — Type

Squeezed(β,ϵ)

Create a Squeezed state with parameters β, ϵ.

SqueezedTwoMode

PhaseSpaceTools.SqueezedTwoMode — Type

SqueezedTwoMode(r,ϕ)

Create a two-mode squeezed state with parameters r, ϕ.

Thermal

PhaseSpaceTools.Thermal — Type

Thermal(β,n̄)

Create a Thermal state with parameters β, n̄.

Bogoliubov

PhaseSpaceTools.Bogoliubov — Type

Bogoliubov(u,v,n̄)

Create a Bogoliubov state with parameters u, v, n̄.