Distributions

The phase-space distributions that may be sampled for these quantum states are

Husimi-Q

PhaseSpaceTools.husimiQ — Function

α,ᾱ = husimiQ(state <: State,N)

Generate N samples from the Husimi-Q phase-space distribution for state.

Moments of the Husimi-Q distribution generate quantum operator averages that are anti-normally ordered.

Glauber-P

PhaseSpaceTools.glauberP — Function

α,α⁺ = glauberP(state <: State,N)

Generate N samples from the Glauber-Sudarshan-P phase-space distribution for state.

Moments of the Glauber-Sudarshan-P distribution generate quantum operator averages that are normally ordered.

Positive-P

PhaseSpaceTools.positiveP — Function

α,α⁺ = positiveP(state <: State,N)

Generate N samples from the positive-P (+P) phase-space distribution for state.

Moments of the +P distribution generate quantum operator averages that are normally ordered.

In general the two random variates α,α⁺ are statistically independent for the +P distribution.

Wigner

PhaseSpaceTools.wigner — Function

α,α⁺ = wigner(state <: State,N)

Generate N samples from wigner phase-space distribution for state.

Moments of the Wigner distribution generate symmetrically ordered quantum operator averages.

Positive-W

PhaseSpaceTools.positiveW — Function

α,α⁺ = positiveW(state <: State,N)

Generate N samples from the positive-W (+W) phase-space distribution for state.

Implemented states are

Fock(N)

The moments of the +W distribution generate quantum operator averages that are symmmetrically ordered.