Distributions

The phase-space distributions that may be sampled for these states are

Husimi-Q

PhaseSpaceTools.husimiQ — Function

α,ᾱ = husimiQ(state <: State,N)

Generate N samples from the Husimi-Q phase-space distribution for state.

Moments of the Husimi-Q distribution generate quantum operator averages that are anti-normally ordered.

Glauber-P

PhaseSpaceTools.glauberP — Function

α,ᾱ = glauberP(state <: State,N)

Generate N samples from the Glauber-Sudarshan-P phase-space distribution for state.

Moments of the Glauber-Sudarshan-P distribution generate quantum operator averages that are normally ordered.

Positive-P

PhaseSpaceTools.positiveP — Function

α,ᾱ = positiveP(state <: State,N)

Generate N samples from the positive-P phase-space distribution for state.

Moments of the positive-P distribution generate quantum operator averages that are normally ordered.

Wigner

PhaseSpaceTools.wigner — Function

α,ᾱ = wigner(state <: State,N)

Generate N samples from wigner phase-space distribution for state.

Moments of the Wigner distribution generate symmetrically ordered quantum operator averages.

Positive-W

PhaseSpaceTools.positiveW — Function

α,ᾱ = positiveW(state <: State,N)

Generate N samples from the positive-W phase-space distribution for state.

Implemented states are

Fock(N)

The moments of the positive-W distribution generate quantum operator averages that are symmmetrically ordered.