Solving SDEs

Noises

randnc

PhaseSpaceTools.randnc — Function

a = randnc(N...)

Return an array of dimension length(N), containing samples of complex random variables with mean zero and variance one:

\[⟨|a|^2⟩ = 1; ⟨a^2⟩ = ⟨(a^*) ^2⟩ = 0,\]

Useful for creating a range of noises that show up in phase-space simulations.

Usage:

a = randnc(10) returns a 10-element Array{Complex{Float64},1}.

a = randnc(50,100) returns a 50x100-element Array{Complex{Float64},2}.

realnoise

PhaseSpaceTools.realnoise — Function

Complex variable SDE's in DifferentialEquations.jl have complex noises by default dispatch on complex fields. This method defines a real noise for use with the library.

realbridge

PhaseSpaceTools.realbridge — Function

Complex variable SDE's in DifferentialEquations.jl have complex noises by default dispatch on complex fields. This method defines a real brownian bridge for use with the library.

Complex noise example

Real noise example

Recovering normal order from Wigner ensemble averages

A common approach to quantum phase space simulations involves performing truncated Wigner simulations, allowing symmetrically ordered operator averages. However, in experiments, the physically measurable moments are usually those in normal order. We can recover normal order for an operator average of a given degree if we have access to all symmetrically ordered operator averages up to the same degree.

One can derive the following useful relationship between normal- and symmetrically-ordered operator averages

\[\langle (a^\dagger)^p a^q\rangle = \sum_{n=0}^{\textrm{min}(p,q)}\frac{(-1)^nn!}{2^n}\binom{p}{n}\binom{q}{n}\langle \{(a^\dagger)^{p-n}a^{q-n}\}_\textrm{sym}\rangle\]

The relationship between normal ordered operator averages and ensemble averages of Wigner trajectories, denoted by $\overline{\left(\dots\right)}_W$ is

\[\langle (a^\dagger)^p a^q\rangle = \sum_{n=0}^{\textrm{min}(p,q)}\frac{(-1)^nn!}{2^n}\binom{p}{n}\binom{q}{n}\overline{\left((\alpha^*)^{p-n}\alpha^{q-n}\right)}_W\]